[refector] 구글 아키텍처 가이드 적용
