+++ app/src/main/java/com/kova700/amazonbookstorepractice/data/api/BookSearchService.kt
@@ -1,7 +1,7 @@
 package com.kova700.amazonbookstorepractice.data.api
 
-import com.kova700.amazonbookstorepractice.data.BookResponse
-import com.kova700.amazonbookstorepractice.feature.main.search.KakaoBookSearchSortType
+import com.kova700.amazonbookstorepractice.data.response.BookResponse
+import com.kova700.amazonbookstorepractice.ui.main.search.KakaoBookSearchSortType
 
 interface BookSearchService {
 
@@ -10,6 +10,6 @@ interface BookSearchService {
 		sort: KakaoBookSearchSortType,
 		page: Int,
 		size: Int
-	) : BookResponse
+	): BookResponse
 
 }
