+++ app/src/main/java/com/kova700/amazonbookstorepractice/data/mapper/BookMapper.kt
@@ -1,0 +1,38 @@
+package com.kova700.amazonbookstorepractice.data.mapper
+
+import com.kova700.amazonbookstorepractice.data.response.NetworkBook
+import com.kova700.amazonbookstorepractice.domain.model.Book
+import kotlinx.collections.immutable.toImmutableList
+
+fun NetworkBook.toDomain() = Book(
+	title = title,
+	thumbnail = thumbnail,
+	authors = authors.toImmutableList(),
+	price = price,
+	contents = contents,
+	datetime = datetime,
+	isbn = isbn,
+	publisher = publisher,
+	status = status,
+	translators = translators.toImmutableList(),
+	url = url,
+	salePrice = salePrice
+)
+
+fun Book.toNetwork() = NetworkBook(
+	title = title,
+	thumbnail = thumbnail,
+	authors = authors.toImmutableList(),
+	price = price,
+	contents = contents,
+	datetime = datetime,
+	isbn = isbn,
+	publisher = publisher,
+	status = status,
+	translators = translators.toImmutableList(),
+	url = url,
+	salePrice = salePrice
+)
+
+fun List<NetworkBook>.toDomain() = this.map { it.toDomain() }
+fun List<Book>.toNetwork() = this.map { it.toNetwork() }
