+++ app/src/main/java/com/kova700/amazonbookstorepractice/domain/repository/BookSearchRepository.kt
@@ -1,6 +1,7 @@
-package com.kova700.amazonbookstorepractice.data
+package com.kova700.amazonbookstorepractice.domain.repository
 
-import com.kova700.amazonbookstorepractice.feature.main.search.KakaoBookSearchSortType
+import com.kova700.amazonbookstorepractice.domain.model.Book
+import com.kova700.amazonbookstorepractice.ui.main.search.KakaoBookSearchSortType
 
 interface BookSearchRepository {
