+++ app/src/main/java/com/kova700/amazonbookstorepractice/ui/main/detail/DetailViewModel.kt
@@ -1,12 +1,10 @@
-package com.kova700.amazonbookstorepractice.feature.main.detail
+package com.kova700.amazonbookstorepractice.ui.main.detail
 
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
-import com.kova700.amazonbookstorepractice.data.Book
-import com.kova700.amazonbookstorepractice.data.BookEntity
-import com.kova700.amazonbookstorepractice.data.BookSearchRepository
-import com.kova700.amazonbookstorepractice.data.toEntity
-import com.kova700.amazonbookstorepractice.feature.main.SELECTED_BOOK_INDEX
+import com.kova700.amazonbookstorepractice.domain.model.Book
+import com.kova700.amazonbookstorepractice.domain.usecase.GetCachedBookUseCase
+import com.kova700.amazonbookstorepractice.ui.main.SELECTED_BOOK_INDEX
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
@@ -15,12 +13,12 @@ import javax.inject.Inject
 @HiltViewModel
 class DetailViewModel @Inject constructor(
 	private val savedStateHandle: SavedStateHandle,
-	private val bookSearchRepository: BookSearchRepository
+	private val getCachedBookUseCase: GetCachedBookUseCase
 ) : ViewModel() {
 
 	private val bookIndex: Int = savedStateHandle.get<Int>(SELECTED_BOOK_INDEX) ?: -1
 
-	private val _viewState = MutableStateFlow<BookEntity>(BookEntity.Default)
+	private val _viewState = MutableStateFlow<Book>(Book.Default)
 	val viewState = _viewState.asStateFlow()
 
 	init {
@@ -28,6 +26,6 @@ class DetailViewModel @Inject constructor(
 	}
 
 	private fun loadBook(index: Int) {
-		_viewState.value = bookSearchRepository.getBook(index).toEntity()
+		_viewState.value = getCachedBookUseCase(index)
 	}
 }
