+++ app/src/main/java/com/kova700/amazonbookstorepractice/ui/main/detail/DetailWebView.kt
@@ -1,4 +1,4 @@
-package com.kova700.amazonbookstorepractice.feature.main.detail
+package com.kova700.amazonbookstorepractice.ui.main.detail
 
 import android.webkit.WebView
 import androidx.compose.runtime.Composable
