+++ app/src/main/java/com/kova700/amazonbookstorepractice/ui/main/search/component/SearchBar.kt
@@ -1,4 +1,4 @@
-package com.kova700.amazonbookstorepractice.feature.main.search.component
+package com.kova700.amazonbookstorepractice.ui.main.search.component
 
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
