+++ app/src/main/java/com/kova700/amazonbookstorepractice/ui/main/search/component/SearchResult.kt
@@ -1,4 +1,4 @@
-package com.kova700.amazonbookstorepractice.feature.main.search.component
+package com.kova700.amazonbookstorepractice.ui.main.search.component
 
 import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
@@ -11,14 +11,14 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
-import com.kova700.amazonbookstorepractice.data.BookEntity
+import com.kova700.amazonbookstorepractice.domain.model.Book
 import kotlinx.collections.immutable.ImmutableList
 
 @Composable
 fun SearchResult(
 	modifier: Modifier = Modifier,
 	onItemClick: (Int) -> Unit = {},
-	books: ImmutableList<BookEntity>,
+	books: ImmutableList<Book>,
 ) {
 	LazyVerticalGrid(
 		columns = GridCells.Fixed(2),
