+++ app/src/main/java/com/kova700/amazonbookstorepractice/ui/main/search/component/SearchResultItem.kt
@@ -1,4 +1,4 @@
-package com.kova700.amazonbookstorepractice.feature.main.search.component
+package com.kova700.amazonbookstorepractice.ui.main.search.component
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
